requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.26.4
python-dotenv==1.0.1
python-dateutil==2.9.0